Split common panel components and utils from shared.js and fixed related errors
