+++ client/src/panels/panel_declarations/covid/covid_expenditures.js
@@ -3,10 +3,10 @@ import React, { Fragment } from "react";
 import MediaQuery from "react-responsive";
 
 
-import {InfographicPanel} from "src/panels/panel_declarations/InfographicPanel.js";
 import {
   declare_panel
-} from "src/panels/panel_declarations/shared.js";
+} from "src/panels/panel_declarations/common_panel_utils.js";
+import {InfographicPanel} from "src/panels/panel_declarations/InfographicPanel.js";
 
 import * as util_components from "src/components/index.js";
