+++ client/src/panels/panel_declarations/covid/covid_intro.js
@@ -1,10 +1,10 @@
 import _ from "lodash";
 import React from "react";
 
-import {InfographicPanel} from "src/panels/panel_declarations/InfographicPanel.js";
 import {
   declare_panel,
-} from "src/panels/panel_declarations/shared.js";
+} from "src/panels/panel_declarations/common_panel_utils.js";
+import {InfographicPanel} from "src/panels/panel_declarations/InfographicPanel.js";
 
 
 import { COVID_EXPENDITUES_FLAG } from "src/models/covid/covid_config.js";
