+++ client/src/panels/panel_declarations/finances/auth_exp_planned_spending/auth_exp_planned_spending.js
@@ -3,11 +3,9 @@ import _ from "lodash";
 import React, { Fragment } from "react";
 
 
+import {declare_panel} from "src/panels/panel_declarations/common_panel_utils.js";
 import {StdPanel, Col} from "src/panels/panel_declarations/InfographicPanel.js";
-import {
-  create_text_maker_component,
-  declare_panel,
-} from "src/panels/panel_declarations/shared.js";
+import {create_text_maker_component} from "src/panels/panel_declarations/shared.js";
 
 
 import * as util_components from "src/components/index.js";
