+++ client/src/panels/panel_declarations/finances/drr_dp_resources/spending_in_perspective.js
@@ -1,11 +1,12 @@
 import _ from "lodash";
 import React from "react";
 
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {InfographicPanel} from "src/panels/panel_declarations/InfographicPanel.js";
 import {
-  declare_panel,
   create_text_maker_component,
 } from "src/panels/panel_declarations/shared.js";
+
 
 import * as util_components from "src/components/index.js";
