+++ client/src/panels/panel_declarations/finances/goco/goco.js
@@ -3,12 +3,14 @@ import _ from "lodash";
 import React, { Fragment } from "react";
 
 
+import { TspanLineWrapper } from "src/panels/panel_declarations/common_panel_components.js";
+import {declare_panel} from "src/panels/panel_declarations/common_panel_utils.js";
 import {InfographicPanel} from "src/panels/panel_declarations/InfographicPanel.js";
 import {
   create_text_maker_component,
-  declare_panel,
-  TspanLineWrapper,
 } from "src/panels/panel_declarations/shared.js";
+
+
 
 import * as util_components from "src/components/index.js";
