+++ client/src/panels/panel_declarations/finances/sobj/personel_spend.js
@@ -1,11 +1,11 @@
 import _ from "lodash";
 import React from "react";
 
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {
   StdPanel,
   Col,
 } from "src/panels/panel_declarations/InfographicPanel.js";
-import { declare_panel } from "src/panels/panel_declarations/shared.js";
 
 import { GraphOverlay } from "src/components/index.js";
