+++ client/src/panels/panel_declarations/finances/sobj/top_spending_areas.js
@@ -2,11 +2,12 @@ import { sum } from "d3-array";
 import _ from "lodash";
 import React from "react";
 
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {StdPanel,Col} from "src/panels/panel_declarations/InfographicPanel.js";
 import {
   create_text_maker_component,
-  declare_panel,
 } from "src/panels/panel_declarations/shared.js";
+
 
 import { is_a11y_mode } from "src/core/injected_build_constants.js";
