+++ client/src/panels/panel_declarations/finances/transfer_payments/historical_g_and_c.js
@@ -1,8 +1,8 @@
 import _ from "lodash";
 import React, { Fragment } from "react";
 
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {InfographicPanel} from "src/panels/panel_declarations/InfographicPanel.js";
-import {declare_panel} from "src/panels/panel_declarations/shared.js";
 
 import * as util_components from "src/components/index.js";
