+++ client/src/panels/panel_declarations/finances/transfer_payments/last_year_g_and_c_perspective.js
@@ -1,8 +1,8 @@
 import _ from "lodash";
 import React, { Fragment } from "react";
 
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {StdPanel, Col} from "src/panels/panel_declarations/InfographicPanel.js";
-import {declare_panel} from "src/panels/panel_declarations/shared.js";
 
 import { is_a11y_mode } from "src/core/injected_build_constants.js";
