+++ client/src/panels/panel_declarations/finances/vote_stat/estimates_in_perspective.js
@@ -1,8 +1,8 @@
 import _ from "lodash";
 import React from "react";
 
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {StdPanel, Col} from "src/panels/panel_declarations/InfographicPanel.js";
-import {declare_panel} from "src/panels/panel_declarations/shared.js";
 
 import FootNote from "src/models/footnotes/footnotes.js";
