+++ client/src/panels/panel_declarations/finances/vote_stat/in_year_estimates_split.js
@@ -3,8 +3,8 @@ import _ from "lodash";
 import React from "react";
 
 
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {StdPanel, Col} from "src/panels/panel_declarations/InfographicPanel.js";
-import {declare_panel} from "src/panels/panel_declarations/shared.js";
 
 
 import { businessConstants } from "src/models/businessConstants.js";
