+++ client/src/panels/panel_declarations/finances/vote_stat/last_year_vote_stat_split.js
@@ -2,8 +2,8 @@ import _ from "lodash";
 import React from "react";
 
 
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {StdPanel, Col} from "src/panels/panel_declarations/InfographicPanel.js";
-import {declare_panel} from "src/panels/panel_declarations/shared.js";
 
 
 import {year_templates} from "src/models/years.js";
