+++ client/src/panels/panel_declarations/finances/welcome_mat/welcome_mat.js
@@ -2,13 +2,16 @@ import classNames from "classnames";
 import _ from "lodash";
 import React, { Fragment } from "react";
 
-import {InfographicPanel} from "src/panels/panel_declarations/InfographicPanel.js";
 import {
-  create_text_maker_component,
   get_planned_fte_source_link,
   get_planned_spending_source_link,
   declare_panel,
+} from "src/panels/panel_declarations/common_panel_utils.js";
+import {InfographicPanel} from "src/panels/panel_declarations/InfographicPanel.js";
+import {
+  create_text_maker_component,
 } from "src/panels/panel_declarations/shared.js";
+
 
 import * as util_components from "src/components/index.js";
