+++ client/src/panels/panel_declarations/misc/gov_related.js
@@ -1,11 +1,12 @@
 import _ from "lodash";
 import React from "react";
 
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {InfographicPanel} from "src/panels/panel_declarations/InfographicPanel.js";
 import {
-  declare_panel,
   TM,
 } from "src/panels/panel_declarations/shared.js";
+
 
 import { create_text_maker } from "src/models/text.js";
