+++ client/src/panels/panel_declarations/misc/hierarchy_panels.js
@@ -1,10 +1,8 @@
 import _ from "lodash";
 import React from "react";
 
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {TextPanel} from "src/panels/panel_declarations/InfographicPanel.js";
-import {
-  declare_panel,
-} from "src/panels/panel_declarations/shared.js";
 
 import * as util_components from "src/components/index.js";
