+++ client/src/panels/panel_declarations/misc/key_concept_panels/key_concept_panels.js
@@ -1,10 +1,9 @@
 import _ from "lodash";
 import React from "react";
 
-import {
-  SomeThingsToKeepInMind,
-  declare_panel,
-} from "src/panels/panel_declarations/shared.js";
+import { SomeThingsToKeepInMind } from "src/panels/panel_declarations/common_panel_components.js";
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
+
 
 import * as util_components from "src/components/index.js";
