+++ client/src/panels/panel_declarations/misc/resource_structure/resource_structure.js
@@ -1,8 +1,8 @@
 import _ from "lodash";
 import React from "react";
 
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {InfographicPanel} from "src/panels/panel_declarations/InfographicPanel.js";
-import {declare_panel} from "src/panels/panel_declarations/shared.js";
 
 
 import { SingleTagResourceExplorer } from "./rooted_resource_scheme.js";
