+++ client/src/panels/panel_declarations/misc/rpb_links.js
@@ -1,11 +1,12 @@
 import _ from "lodash";
 import React from "react";
 
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {TextPanel} from "src/panels/panel_declarations/InfographicPanel.js";
 import {
   create_text_maker_component,
-  declare_panel,
 } from "src/panels/panel_declarations/shared.js";
+
 
 import * as util_components from "src/components/index.js";
