+++ client/src/panels/panel_declarations/misc/tags_related_to_subject_panels.js
@@ -2,11 +2,12 @@ import classNames from "classnames";
 import _ from "lodash";
 import React, { Fragment } from "react";
 
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {TextPanel} from "src/panels/panel_declarations/InfographicPanel.js";
 import {
   create_text_maker_component,
-  declare_panel,
 } from "src/panels/panel_declarations/shared.js";
+
 
 import * as util_components from "src/components/index.js";
