+++ client/src/panels/panel_declarations/misc/warning_panels.js
@@ -1,10 +1,11 @@
 import _ from "lodash";
 import React, { Fragment } from "react";
 
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {
   create_text_maker_component,
-  declare_panel,
 } from "src/panels/panel_declarations/shared.js";
+
 
 import * as util_components from "src/components/index.js";
