+++ client/src/panels/panel_declarations/people/employee_age.js
@@ -2,6 +2,7 @@ import { sum } from "d3-array";
 import _ from "lodash";
 import React from "react";
 
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {
   StdPanel,
   Col,
@@ -9,7 +10,6 @@ import {
 import {
   TabbedContent,
   create_text_maker_component,
-  declare_panel,
 } from "src/panels/panel_declarations/shared.js";
 
 import { GraphOverlay } from "src/components/index.js";
