+++ client/src/panels/panel_declarations/people/employee_executive_level.js
@@ -2,11 +2,12 @@ import { sum } from "d3-array";
 import _ from "lodash";
 import React from "react";
 
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {StdPanel, Col} from "src/panels/panel_declarations/InfographicPanel.js";
 import {
-  create_text_maker_component,
-  declare_panel,
+  create_text_maker_component
 } from "src/panels/panel_declarations/shared.js";
+
 
 import { businessConstants } from "src/models/businessConstants.js";
 import {run_template} from "src/models/text.js";
