+++ client/src/panels/panel_declarations/people/employee_last_year_totals.js
@@ -1,11 +1,12 @@
 import React from "react";
 
 
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {StdPanel, Col} from "src/panels/panel_declarations/InfographicPanel.js";
 import {
-  create_text_maker_component,
-  declare_panel,
+  create_text_maker_component
 } from "src/panels/panel_declarations/shared.js";
+
 
 
 import {year_templates} from "src/models/years.js";
