+++ client/src/panels/panel_declarations/results/drr_summary.js
@@ -3,8 +3,8 @@ import _ from "lodash";
 import React, { Fragment } from "react";
 
 
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {InfographicPanel} from "src/panels/panel_declarations/InfographicPanel.js";
-import {declare_panel} from "src/panels/panel_declarations/shared.js";
 
 
 import * as util_components from "src/components/index.js";
