+++ client/src/panels/panel_declarations/results/gov_drr.js
@@ -1,11 +1,10 @@
 import _ from "lodash";
 import React from "react";
 
+import {HeightClippedGraph} from "src/panels/panel_declarations/common_panel_components.js";
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {InfographicPanel} from "src/panels/panel_declarations/InfographicPanel.js";
-import {
-  declare_panel,
-  HeightClippedGraph,
-} from "src/panels/panel_declarations/shared.js";
+
 
 import {
   SmartDisplayTable,
