+++ client/src/panels/panel_declarations/results/results_intro.js
@@ -2,11 +2,12 @@ import { sum } from "d3-array";
 import _ from "lodash";
 import React from "react";
 
+import { declare_panel } from "src/panels/panel_declarations/common_panel_utils.js";
 import {InfographicPanel} from "src/panels/panel_declarations/InfographicPanel.js";
 import {
-  declare_panel,
-  create_text_maker_component,
+  create_text_maker_component
 } from "src/panels/panel_declarations/shared.js";
+
 
 import { Subject } from "src/models/subject.js";
